wip: extract common behavior
